requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
shap
joblib
jupyter
notebook
ipykernel
sqlalchemy
psycopg2-binary
pyyaml
pyarrow
pytest
torch
implicit
scipy
kagglehub
great_expectations
boto3
